structure finish

+++ aust-pojo/src/main/java/cn/edu/aust/pojo/entity/SolutionDO.java
@@ -34,7 +34,7 @@ public class SolutionDO {
     private String language;
 
     @Column(name = "contest_id")
-    private Integer contestId;
+    private Long contestId;
 
     private Integer verdict;
